+++ src/partials/home.html
@@ -61,24 +61,58 @@
         <p class="footer-text"> Developed with
             <svg class="footer-icon" width="14" height="12">
                 <use href="./images/symbol-defs.svg#icon-heart"></use>
-            </svg> by
+            </svg> 
             <a class="footer-link" href="" data-modal-open>GoIt Students</a>
         </p>
         <div  data-modal class="backdrop is-hidden">
             <div class="our__team-modal">
                 <h3 class="our_team-title">Our team</h3>
                 <ul class = "team_list">
-                    <li class="team_item"></li>
-                    <li class="team_item"></li>
-                    <li class="team_item"></li>
-                    <li class="team_item"></li>
-                    <li class="team_item"></li>
-                    <li class="team_item"></li>
-                    <li class="team_item"></li>
-                    <li class="team_item"></li>
-                    <li class="team_item"></li>
+                   <li class="team_item">
+                        <img class="team_img"  src="https://img.kupigolos.ru/hero/5cdb2ce28b1e4.png?p=bh&s=0aa83e7d3d9cac880eb4620efc523870" alt="img">
+                        <p class="team_disc">Kostiantyn Oliinyk</p>
+                    </li>
+                    <li class="team_item">
+                        <img class="team_img" src="https://i.pinimg.com/236x/bf/e4/64/bfe46464a535dc53b48fc925e4304433--agnes-despicable-me-cartoon-movies.jpg" alt="img">
+                        <p class="team_disc">Olesya LZ</p>
+                    </li>
+                    <li class="team_item">
+                        <img class="team_img" src="https://i.pinimg.com/236x/82/37/5f/82375fa0115f41490b1129ad44794071--funny-minion-minions-quotes.jpg"
+                            alt="img">
+                        <p class="team_disc">Antonina Bondar</p>
+                    </li>
+                    <li class="team_item">
+                        <img class="team_img" src="https://i.pinimg.com/236x/82/37/5f/82375fa0115f41490b1129ad44794071--funny-minion-minions-quotes.jpg"
+                            alt="img">
+                        <p class="team_disc">Oksana ???</p>
+                    </li>
+                    <li class="team_item">
+                        <img class="team_img" src="https://i.pinimg.com/236x/82/37/5f/82375fa0115f41490b1129ad44794071--funny-minion-minions-quotes.jpg"
+                            alt="img">
+                        <p class="team_disc">Oleksii Putnenko</p>
+                    </li>
+                    <li class="team_item">
+                        <img class="team_img" src="https://i.pinimg.com/236x/82/37/5f/82375fa0115f41490b1129ad44794071--funny-minion-minions-quotes.jpg"
+                            alt="img">
+                        <p class="team_disc">Taras Shapovalov</p>
+                    </li>
+                    <li class="team_item">
+                        <img class="team_img" src="https://i.pinimg.com/236x/82/37/5f/82375fa0115f41490b1129ad44794071--funny-minion-minions-quotes.jpg"
+                            alt="img">
+                        <p class="team_disc">Yuliia Kovalchuk</p>
+                    </li>
+                    <li class="team_item">
+                        <img class="team_img" src="https://i.pinimg.com/236x/82/37/5f/82375fa0115f41490b1129ad44794071--funny-minion-minions-quotes.jpg"
+                            alt="img">
+                        <p class="team_disc">Alex Sidorenko</p>
+                    </li>
+                    <li class="team_item">
+                        <img class="team_img" src="https://i.pinimg.com/236x/82/37/5f/82375fa0115f41490b1129ad44794071--funny-minion-minions-quotes.jpg"
+                            alt="img">
+                        <p class="team_disc">Juliia Volokhovych</p>
+                    </li>
                 </ul>
-                <button type="button" data-modal-close class="close__modal">close modal</button>
+                <button type="button" data-modal-close class="close__modal"><img class="modal-svg" src="./images/modal-team/svg-close.svg"></button>
             </div>
         </div>
     </div>
