+++ src/sass/_backdrop.scss
@@ -39,10 +39,29 @@
     position: absolute;
     right: 20px;
     top:20px;
+    width: 40px;
+    height: 40px;
+    border-radius:50% ;
+    border:none;
+    justify-content: center;
+    align-items: center;
+    background-color: transparent;
+}
+.close__modal:hover,
+.close__modal:focus{
+    background-color:#ff6b01;
+}
+.modal-svg{
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%,-50%);
+    height: 35px;
+    width: 35px;
 }
 .our_team-title{
     font-size: 30px;
-    color: red;
+    color: #ff6b01;
    text-align: center;
    margin-top:30px;
 }
@@ -58,7 +77,46 @@
     list-style: none;
     width: 160px;
     height: 120px;
-    background-color: yellow;
+  border: 1px solid #000;
     margin-right: 5px;
     margin-bottom: 5px;
+    overflow: hidden;
+    position: relative;
+    color: #dd630b;
+}
+.team_item:hover,
+.team_item:focus {
+overflow: visible;
+}
+.team_item:hover>.team_img,
+.team_item:focus>.team_img
+{
+    z-index: 1000;
+    position: absolute;
+    width: 120%;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+ 
+border: 1px solid #dd630b;
+color: #9b4406;
+}
+.team_img{
+
+    display: block;
+   width: 100%;
+    -webkit-transition: 
+        width 500ms cubic-bezier(0.4, 0, 0.2, 1) ;
+    transition: 
+        width 500ms cubic-bezier(0.4, 0, 0.2, 1);
+ 
+}
+.team_disc{
+display: block;
+position: absolute;
+bottom: 5px;
+left: 5px;
+font-size: 20px;
+font-weight: 500;
+
 }
